server/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg[binary]==3.2.13
gunicorn==23.0.0
